refactor: 减少头文件引用以加快编译速度 1
