+++ src/Magpie.Core/ExportHelper.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#ifdef MAGPIE_CORE_EXPORTS
+#define API_DECLSPEC __declspec(dllexport)
+#else
+#define API_DECLSPEC __declspec(dllimport)
+#endif
