+++ src/Magpie.Core/MagOptions.h
@@ -1,7 +1,5 @@
 #pragma once
-#include "pch.h"
 #include <parallel_hashmap/phmap.h>
-
 
 namespace Magpie::Core {
