+++ src/Magpie.UI/EffectsService.cpp
@@ -4,9 +4,9 @@
 #include "StrUtils.h"
 #include "Win32Utils.h"
 #include "CommonSharedConstants.h"
+#include <d3dcompiler.h>	// ID3DBlob
 
 using namespace Magpie::Core;
-
 
 namespace winrt::Magpie::UI {
