+++ src/Magpie.UI/ScalingProfile.h
@@ -2,7 +2,6 @@
 #include "pch.h"
 #include <Magpie.Core.h>
 
-
 namespace winrt::Magpie::UI {
 
 enum class CursorScaling {
